moved the floating bubbles to the Home Screen
